requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.19.0
attrs==23.2.0
certifi==2023.11.17
chardet==3.0.4
discord.py==2.3.2
frozenlist==1.4.1
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==6.0.4
python-dotenv==1.0.0
rfc3986==1.5.0
sniffio==1.3.0
yarl==1.9.4
